Consolidate ContentInfo classes
There were two ContentInfo classes, which differed only by a single
whitespace character!

One copy was raised up the hierarchy to a more common location and the
other was deleted.
