+++ base/src/main/java/org/mozilla/jss/pkix/cmc/EncryptedPOP.java
@@ -40,13 +40,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
+import org.mozilla.jss.ContentInfo;
 import org.mozilla.jss.asn1.ASN1Template;
 import org.mozilla.jss.asn1.ASN1Value;
 import org.mozilla.jss.asn1.InvalidBERException;
 import org.mozilla.jss.asn1.OCTET_STRING;
 import org.mozilla.jss.asn1.SEQUENCE;
 import org.mozilla.jss.asn1.Tag;
-import org.mozilla.jss.pkix.cms.ContentInfo;
 import org.mozilla.jss.pkix.primitive.AlgorithmIdentifier;
 
 /**
