+++ base/src/main/java/org/mozilla/jss/pkix/cmc/TaggedContentInfo.java
@@ -8,13 +8,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
+import org.mozilla.jss.ContentInfo;
 import org.mozilla.jss.asn1.ASN1Template;
 import org.mozilla.jss.asn1.ASN1Value;
 import org.mozilla.jss.asn1.INTEGER;
 import org.mozilla.jss.asn1.InvalidBERException;
 import org.mozilla.jss.asn1.SEQUENCE;
 import org.mozilla.jss.asn1.Tag;
-import org.mozilla.jss.pkix.cms.ContentInfo;
 
 /**
  * CMC <i>TaggedContentInfo</i>.
