+++ base/src/main/java/org/mozilla/jss/pkix/cms/EncryptedContentInfo.java
@@ -14,6 +14,7 @@ import java.security.spec.AlgorithmParameterSpec;
 
 import javax.crypto.BadPaddingException;
 
+import org.mozilla.jss.ContentInfo;
 import org.mozilla.jss.CryptoManager;
 import org.mozilla.jss.NotInitializedException;
 import org.mozilla.jss.asn1.ASN1Template;
