+++ base/src/main/java/org/mozilla/jss/pkix/cms/SignerInfo.java
@@ -13,6 +13,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.PublicKey;
 import java.security.SignatureException;
 
+import org.mozilla.jss.ContentInfo;
 import org.mozilla.jss.CryptoManager;
 import org.mozilla.jss.NotInitializedException;
 import org.mozilla.jss.asn1.ANY;
